pyproject.toml
[project]
name = "lung-cancer-screening"
version = "0.1.0"
description = ""
authors = [{name = "Your Name",email = "you@example.com"}]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "azure-identity (>=1.23.0,<2.0.0)",
    "django (>=5.2.4,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "nhsuk-frontend-jinja (>=0.4.1,<0.5.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "mozilla-django-oidc (>=3.0.0,<4.0.0)",
    "pyjwt (>=2.8.0,<3.0.0)",
    "cryptography (>=41.0.0,<43.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.7"
playwright = "^1.56.0"
axe-playwright-python = "^0.1.7"
behave-django = "^1.8.0"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.behave]
paths = ["tests/features"]
